microservices/others/gps_updater/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
redis==5.2.0